refactor: overwrite old startCQCLog.py
